Create InputReady CursorTrigger (BEAM)
The macro makes sure that input field is accessible (e.g. in Firefox) before executing code further. As a plugable block it can replace "wait for mouse click", "wait for page to load", "if window on top" commands, so instead of clicking you just move mouse cursor to the desired field. It is especially useful in web-browsers, where "get control" command doesn't work.